% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\name{nascSynth}
\alias{nascSynth}
\title{Bayesian Network-Aware Synthetic Control}
\description{
An R6 class implementing Bayesian Synthetic Control estimators. Supports
single and multiple treated units, optional covariate adjustment, and
predictor matching. Stan models are selected automatically based on the
structure of the supplied data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-nascSynth-new}{\code{nascSynth$new()}}
\item \href{#method-nascSynth-fit}{\code{nascSynth$fit()}}
\item \href{#method-nascSynth-updateWidth}{\code{nascSynth$updateWidth()}}
\item \href{#method-nascSynth-summarizeLift}{\code{nascSynth$summarizeLift()}}
\item \href{#method-nascSynth-effectPlot}{\code{nascSynth$effectPlot()}}
\item \href{#method-nascSynth-clone}{\code{nascSynth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-new"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$new(
  data,
  time,
  id,
  treated,
  outcome,
  ci_width = 0.75,
  covariates0 = NULL,
  covariates1 = NULL,
  vs = NULL,
  W = NULL,
  spatial_model = "none"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-fit"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$fit(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-updateWidth"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-updateWidth}{}}}
\subsection{Method \code{updateWidth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$updateWidth(ci_width = 0.75)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-summarizeLift"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-summarizeLift}{}}}
\subsection{Method \code{summarizeLift()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$summarizeLift()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-effectPlot"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-effectPlot}{}}}
\subsection{Method \code{effectPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$effectPlot(facet = TRUE, subset = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nascSynth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-nascSynth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{nascSynth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
